refactor: extracting google button into a separate component
