+++ frontend/app/(login)/login/LoginForm.tsx
@@ -3,15 +3,14 @@
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useRouter } from 'next/navigation';
 import axios, { AxiosError } from 'axios';
-import { signIn } from 'next-auth/react';
 import { useState } from "react";
 import Link from "next/link";
 
 import ShowPassword from '@/app/components/ShowPassword';
 import ReturnLogo from "@/app/logos-icons/Return";
-import GoogleIcon from "@/app/logos-icons/Google";
 import BSWImage from "@/app/components/BSWImage";
 import { useUser } from '@/app/hooks/userHook';
+import GoogleOAuth from '@/app/components/GoogleOAuth';
 
 interface LoginFormInterface {
   email: string,
@@ -48,10 +47,6 @@ const LoginForm = () => {
         console.log(error)
       }
     }
-  }
-
-  const googleSignIn = async () => {
-    await signIn('google', { redirectTo: '/' })
   }
 
   return (  
@@ -115,10 +110,7 @@ const LoginForm = () => {
                       Recuperar contraseña
                     </Link>
                   </button>
-                  <button onClick={googleSignIn} className="flex items-center justify-center space-x-2 w-full min-w-[330px] 2xl:w-1/2 py-2.5 text-xl font-large text-center text-white border border-blue-900 transition duration-200 ease-in-out transform bg-trasparent hover:border hover:border-blue-700 hover:bg-blue-800">
-                    <GoogleIcon width="25" height="25" />
-                    <span>Iniciar sesión con Google</span>
-                  </button>
+                  <GoogleOAuth />
                 </span>
               </section>
             </form>
